increase and move location of memory in OCRAM linker file
Move the memory (m_data) area in the OCRAM linker file to the
EPDC OCRAM area. Also use the code bus for the code area instead
of the system bus which should improve performance. This allows
firmware of up to ~64KiB and memory of 128KiB.

Note: To get more heap space in FreeRTOS configTOTAL_HEAP_SIZE in
FreeRTOSConfig.h needs to be adjusted as well.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>