restore stackpointer in vector table
When using the elf file format the vector table gets linked directly
to 0x0 (0x180000 for data bus/A7 bus). U-Boots bootaux command tries
to be helpful and writes the start address to that location (since
this was necessary for binary files which have a relocated vector
table). This is not usually a problem, since we restore the stack
pointer imeaditly at the beginning.

However, the FreeRTOS port code relies on a complete vector table
(see prvPortStartFirstTask in Source/portable/GCC/ARM_CM4F/port.c).

This needs to be fixed in U-Boot: U-Boot should not write into the
vector table if it is provided by the elf binary.

For now, work around by restoring the stack pointer entry in the
vector table in the firmware startup code.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>